+++ engine/flink/test-utils/src/main/scala/pl/touk/nussknacker/engine/process/helpers/SampleNodes.scala
@@ -754,10 +754,12 @@ object SampleNodes {
     private val aTypeDeclaration = ParameterDeclaration
       .mandatory[String](aTypeParamName)
       .withCreator(modify =
-        _.copy(editor =
+        _.copy(editors =
           Some(
-            FixedValuesParameterEditor(
-              List(FixedExpressionValue("'type1'", "type1"), FixedExpressionValue("'type2'", "type2"))
+            ParameterEditors(
+              FixedValuesParameterEditor(
+                List(FixedExpressionValue("'type1'", "type1"), FixedExpressionValue("'type2'", "type2"))
+              )
             )
           )
         )
@@ -769,8 +771,12 @@ object SampleNodes {
       .mandatory[Int](versionParamName)
       .withAdvancedCreator[List[Int]](
         create = versions =>
-          _.copy(editor =
-            Some(FixedValuesParameterEditor(versions.map(v => FixedExpressionValue(v.toString, v.toString))))
+          _.copy(editors =
+            Some(
+              ParameterEditors(
+                FixedValuesParameterEditor(versions.map(v => FixedExpressionValue(v.toString, v.toString)))
+              )
+            )
           )
       )
 
@@ -924,10 +930,12 @@ object SampleNodes {
     private val aTypeParamDeclaration = ParameterDeclaration
       .mandatory[String](aTypeParamName)
       .withCreator(
-        modify = _.copy(editor =
+        modify = _.copy(editors =
           Some(
-            FixedValuesParameterEditor(
-              List(FixedExpressionValue("'type1'", "type1"), FixedExpressionValue("'type2'", "type2"))
+            ParameterEditors(
+              FixedValuesParameterEditor(
+                List(FixedExpressionValue("'type1'", "type1"), FixedExpressionValue("'type2'", "type2"))
+              )
             )
           )
         )
@@ -939,8 +947,12 @@ object SampleNodes {
       .mandatory[Int](versionParamName)
       .withAdvancedCreator[List[Int]](
         create = versions =>
-          _.copy(editor =
-            Some(FixedValuesParameterEditor(versions.map(v => FixedExpressionValue(v.toString, v.toString))))
+          _.copy(editors =
+            Some(
+              ParameterEditors(
+                FixedValuesParameterEditor(versions.map(v => FixedExpressionValue(v.toString, v.toString)))
+              )
+            )
           )
       )