+++ scenario-compiler/src/test/scala/pl/touk/nussknacker/engine/definition/component/parameter/editor/EditorExtractorTest.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.engine.definition.component.parameter.editor
 
+import cats.implicits.catsSyntaxOptionId
 import com.cronutils.model.Cron
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
@@ -20,20 +21,18 @@ class EditorExtractorTest extends AnyFunSuite with Matchers {
   private def notAnnotated(param: String) = ()
 
   private def dualEditorAnnotated(
-      @DualEditor(
-        simpleEditor = new SimpleEditor(
-          `type` = SimpleEditorType.FIXED_VALUES_EDITOR,
-          possibleValues = Array(new LabeledExpression(expression = "'test'", label = "test2"))
-        ),
-        defaultMode = DualEditorMode.SIMPLE
-      ) param: String
+      @SimpleEditor(
+        `type` = SimpleEditorType.FIXED_VALUES_EDITOR,
+        possibleValues = Array(new LabeledExpression(expression = "'test'", label = "test2")),
+      )
+      @SpelEditor
+      param: String
   ) = ()
 
   private def dualEditorAnnotatedLazy(
-      @DualEditor(
-        simpleEditor = new SimpleEditor(`type` = SimpleEditorType.DATE_EDITOR),
-        defaultMode = DualEditorMode.SIMPLE
-      ) param: LazyParameter[String]
+      @SimpleEditor(`type` = SimpleEditorType.DATE_EDITOR)
+      @SpelEditor
+      param: LazyParameter[String]
   ) = ()
 
   private def simpleEditorAnnotated(@SimpleEditor(`type` = SimpleEditorType.BOOL_EDITOR) param: String) = ()
@@ -42,9 +41,9 @@ class EditorExtractorTest extends AnyFunSuite with Matchers {
       @SimpleEditor(`type` = SimpleEditorType.BOOL_EDITOR) param: LazyParameter[String]
   ) = ()
 
-  private def rawEditorAnnotated(@RawEditor param: String) = ()
+  private def rawEditorAnnotated(@SpelEditor param: String) = ()
 
-  private def rawEditorAnnotatedLazy(@RawEditor param: LazyParameter[String]) = ()
+  private def rawEditorAnnotatedLazy(@SpelEditor param: LazyParameter[String]) = ()
 
   private def simpleParams(
       javaEnum: JavaSampleEnum,
@@ -71,158 +70,145 @@ class EditorExtractorTest extends AnyFunSuite with Matchers {
   private val paramRawEditorAnnotatedLazy = getFirstParam("rawEditorAnnotatedLazy", classOf[LazyParameter[String]])
 
   test("assign RawEditor when no annotation detected") {
-    EditorExtractor.extract(paramNotAnnotated, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(StringParameterEditor, DualEditorMode.RAW)
-    )
+    EditorExtractor.extract(paramNotAnnotated, ParameterConfig.empty) shouldBe ParameterEditors(
+      SpelParameterEditor,
+      SpelTemplateParameterEditor,
+    ).some
   }
 
-  test("detect @DualEditor annotation") {
+  test("detect either @SimpleEditor and @SpelEditor annotations") {
 
     EditorExtractor.extract(paramDualEditorAnnotated, ParameterConfig.empty) shouldBe
-      Some(
-        DualParameterEditor(
-          simpleEditor = FixedValuesParameterEditor(
-            possibleValues = List(FixedExpressionValue("'test'", "test2"))
-          ),
-          defaultMode = DualEditorMode.SIMPLE
-        )
-      )
+      ParameterEditors(
+        FixedValuesParameterEditor(
+          possibleValues = List(FixedExpressionValue("'test'", "test2"))
+        ),
+        SpelParameterEditor,
+      ).some
 
     EditorExtractor.extract(paramDualEditorLazyAnnotated, ParameterConfig.empty) shouldBe
-      Some(
-        DualParameterEditor(
-          simpleEditor = DateParameterEditor,
-          defaultMode = DualEditorMode.SIMPLE
-        )
-      )
+      ParameterEditors(
+        DateParameterEditor,
+        SpelParameterEditor,
+      ).some
   }
 
   test("detect @SimpleEditor annotation") {
 
     EditorExtractor.extract(paramSimpleEditorAnnotated, ParameterConfig.empty) shouldBe
-      Some(BoolParameterEditor)
+      ParameterEditors(BoolParameterEditor).some
 
     EditorExtractor.extract(paramSimpleEditorLazyAnnotated, ParameterConfig.empty) shouldBe
-      Some(BoolParameterEditor)
+      ParameterEditors(BoolParameterEditor).some
   }
 
-  test("detect @RawEditor annotation") {
-    EditorExtractor.extract(paramRawEditorAnnotated, ParameterConfig.empty) shouldBe Some(RawParameterEditor)
-    EditorExtractor.extract(paramRawEditorAnnotatedLazy, ParameterConfig.empty) shouldBe Some(RawParameterEditor)
+  test("detect @SpelEditor annotation") {
+    EditorExtractor.extract(paramRawEditorAnnotated, ParameterConfig.empty) shouldBe ParameterEditors(
+      SpelParameterEditor
+    ).some
+    EditorExtractor.extract(paramRawEditorAnnotatedLazy, ParameterConfig.empty) shouldBe ParameterEditors(
+      SpelParameterEditor
+    ).some
   }
 
   test("determine editor by config") {
     val fixedValuesEditor = FixedValuesParameterEditor(List(FixedExpressionValue("'expression'", "label")))
-    val config            = ParameterConfig(None, Some(fixedValuesEditor), None, None, None)
+    val config            = ParameterConfig(None, Some(List(fixedValuesEditor)), None, None, None)
 
-    EditorExtractor.extract(paramNotAnnotated, config) shouldBe Some(fixedValuesEditor)
+    EditorExtractor.extract(paramNotAnnotated, config) shouldBe ParameterEditors(fixedValuesEditor).some
   }
 
   test("determine editor by type enum") {
     val param = getSimpleParamByType[JavaSampleEnum]
 
-    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(
-        FixedValuesParameterEditor(
-          List(
-            FixedExpressionValue(
-              s"T(${classOf[JavaSampleEnum].getName}).${JavaSampleEnum.FIRST_VALUE.name()}",
-              "first_value"
-            ),
-            FixedExpressionValue(
-              s"T(${classOf[JavaSampleEnum].getName}).${JavaSampleEnum.SECOND_VALUE.name()}",
-              "second_value"
-            )
+    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe ParameterEditors(
+      FixedValuesParameterEditor(
+        List(
+          FixedExpressionValue(
+            s"T(${classOf[JavaSampleEnum].getName}).${JavaSampleEnum.FIRST_VALUE.name()}",
+            "first_value"
+          ),
+          FixedExpressionValue(
+            s"T(${classOf[JavaSampleEnum].getName}).${JavaSampleEnum.SECOND_VALUE.name()}",
+            "second_value"
           )
-        ),
-        DualEditorMode.SIMPLE
-      )
-    )
+        )
+      ),
+      SpelParameterEditor
+    ).some
   }
 
   test("determine editor by type LocalDateTime") {
     val param = getSimpleParamByType[LocalDateTime]
 
-    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(
-        simpleEditor = DateTimeParameterEditor,
-        defaultMode = DualEditorMode.SIMPLE
-      )
-    )
+    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe ParameterEditors(
+      DateTimeParameterEditor,
+      SpelParameterEditor,
+    ).some
   }
 
   test("determine editor by type LocalDate") {
     val param = getSimpleParamByType[LocalDate]
 
-    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(
-        simpleEditor = DateParameterEditor,
-        defaultMode = DualEditorMode.SIMPLE
-      )
-    )
+    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe ParameterEditors(
+      DateParameterEditor,
+      SpelParameterEditor,
+    ).some
   }
 
   test("determine editor by type LocalTime") {
     val param = getSimpleParamByType[LocalTime]
 
-    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(
-        simpleEditor = TimeParameterEditor,
-        defaultMode = DualEditorMode.SIMPLE
-      )
-    )
+    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe ParameterEditors(
+      TimeParameterEditor,
+      SpelParameterEditor,
+    ).some
   }
 
   test("determine editor by type Duration") {
     val param = getSimpleParamByType[Duration]
 
-    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(
-        simpleEditor = DurationParameterEditor(List(ChronoUnit.DAYS, ChronoUnit.HOURS, ChronoUnit.MINUTES)),
-        defaultMode = DualEditorMode.SIMPLE
-      )
-    )
+    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe ParameterEditors(
+      DurationParameterEditor(List(ChronoUnit.DAYS, ChronoUnit.HOURS, ChronoUnit.MINUTES)),
+      SpelParameterEditor,
+    ).some
   }
 
   test("determine editor by config for Duration") {
     val param  = getSimpleParamByType[Duration]
     val editor = DurationParameterEditor(timeRangeComponents = List(ChronoUnit.MINUTES))
 
-    EditorExtractor.extract(param, ParameterConfig.empty.copy(editor = Some(editor))) shouldBe Some(editor)
+    EditorExtractor.extract(param, ParameterConfig.empty.copy(editors = Some(List(editor)))) shouldBe ParameterEditors(
+      editor
+    ).some
   }
 
   test("determine editor by type Period") {
     val param = getSimpleParamByType[Period]
 
-    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(
-        simpleEditor = PeriodParameterEditor(List(ChronoUnit.YEARS, ChronoUnit.MONTHS, ChronoUnit.DAYS)),
-        defaultMode = DualEditorMode.SIMPLE
-      )
-    )
+    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe ParameterEditors(
+      PeriodParameterEditor(List(ChronoUnit.YEARS, ChronoUnit.MONTHS, ChronoUnit.DAYS)),
+      SpelParameterEditor,
+    ).some
   }
 
   test("determine editor by type Cron") {
     val param = getSimpleParamByType[Cron]
 
-    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe Some(
-      DualParameterEditor(
-        simpleEditor = CronParameterEditor,
-        defaultMode = DualEditorMode.SIMPLE
-      )
-    )
+    EditorExtractor.extract(param, ParameterConfig.empty) shouldBe ParameterEditors(
+      CronParameterEditor,
+      SpelParameterEditor,
+    ).some
   }
 
   test("determine editor by type Charsequence") {
     val charseqParam = getSimpleParamByType[CharSequence]
     val stringParam  = getSimpleParamByType[String]
 
-    val expectedEditor = Some(
-      DualParameterEditor(
-        simpleEditor = StringParameterEditor,
-        defaultMode = DualEditorMode.RAW
-      )
-    )
+    val expectedEditor = ParameterEditors(
+      SpelParameterEditor,
+      SpelTemplateParameterEditor,
+    ).some
     EditorExtractor.extract(charseqParam, ParameterConfig.empty) shouldBe expectedEditor
     EditorExtractor.extract(stringParam, ParameterConfig.empty) shouldBe expectedEditor
   }